:sparkles: remove tabs and add comments
