+++ app/[usernameSlug]/page.tsx
@@ -8,7 +8,7 @@ import { ProfileHeader } from './profile-header'
 import { SignOutButton } from './sign-out-button'
 import { getFullTxnsByUser, getTxnsByUser } from '@/db/txn'
 import { getProjectsByUser } from '@/db/project'
-import { ProfileTabs } from './profile-tabs'
+import { ProfileHistory } from './profile-history'
 import { getBidsByUser } from '@/db/bid'
 import { getCommentsByUser } from '@/db/comment'
 
@@ -36,24 +36,22 @@ export default async function UserProfilePage(props: {
   ])
   const isOwnProfile = user?.id === profile?.id
   return (
-    <div className="flex flex-col p-3 sm:p-5">
+    <div className="flex flex-col gap-8 p-3 sm:p-5">
       <ProfileHeader profile={profile} isOwnProfile={isOwnProfile} />
-      <div className="flex flex-col gap-10">
-        <ProfileTabs
-          profile={profile}
-          projects={projects}
-          comments={comments}
-          bids={bids}
-          txns={txns}
-          userProfile={userProfile}
-          userTxns={userTxns}
-        />
-        {isOwnProfile && (
-          <div className="mt-5 flex justify-center">
-            <SignOutButton />
-          </div>
-        )}
-      </div>
+      <ProfileHistory
+        profile={profile}
+        projects={projects}
+        comments={comments}
+        bids={bids}
+        txns={txns}
+        userProfile={userProfile}
+        userTxns={userTxns}
+      />
+      {isOwnProfile && (
+        <div className="mt-5 flex justify-center">
+          <SignOutButton />
+        </div>
+      )}
     </div>
   )
 }
