+++ app/[usernameSlug]/profile-investments.tsx
